bring forward the sv_extentions and stdver changes from 0.3

+++ qw/include/client.h
@@ -255,7 +255,7 @@ typedef struct
 	char		levelname[40];	// for display on solo scoreboard
 	int			playernum;
 	int			viewentity;
-	int			stdver;
+	float		stdver;
 
 // refresh related state
 	struct model_s	*worldmodel;	// cl_entitites[0].model
